+++ src/app.ts
@@ -3,7 +3,7 @@ import express from "express";
 import { healthRouter, userRouter, questRouter, tagRouter, weeklyReportRouter, badgeRouter } from "./route";
 import { cookie } from "express-validator";
 import cookieParser from "cookie-parser";
-import { logger } from "./pkg/logger";
+import { logger } from "./pkg/logger/logger";
 import { requestsLogger } from "./route/middlewares/requestsLogger";
 import { badgeCronJob, weeklyReportCronJob, questCronJob } from "./cron-job";
 import helmet from "helmet";
