+++ src/controller/badge/badge_controller.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from "express";
-import { getUserIdFromToken } from "../../pkg/jwt";
+import { getUserIdFromToken } from "../../pkg/jwt/jwt";
 import { HttpError } from "../../utils/httpError";
 import { achieveBadgeService, listBadgesService, pinBadgeService, unpinBadgeService } from "../../service/badge";
 import { AchieveBadgeResponse, ListBadgesResponse, PinBadgeResponse, UnpinBadgeResponse } from "./response";
