+++ src/controller/tag/tag_controller.ts
@@ -1,7 +1,7 @@
 import { Request, Response } from "express";
 import { CreateTagRequest, UpdateTagRequest } from "./request";
 import { CreateTagResponse, UpdateTagResponse, DeleteTagResponse, ListTagsResponse } from "./response";
-import { getUserIdFromToken } from "../../pkg/jwt";
+import { getUserIdFromToken } from "../../pkg/jwt/jwt";
 import { HttpError } from "../../utils/httpError";
 import { Tag } from "../../model/tag/types";
 import { createTagService, deleteTagService, listTagsService, updateTagService } from "../../service/tag";
