+++ src/cron-job/weeklyReport.ts
@@ -1,5 +1,5 @@
 import cron from "node-cron";
-import { logger } from "../pkg/logger";
+import { logger } from "../pkg/logger/logger";
 import { prisma } from "../db/db";
 import { WeeklyReportModel } from "../model/weeklyReport/weekly_report_model";
 import { getStartAndEndUTCDateTime } from "../utils/datetime";
