+++ src/pkg/firebase/config.ts
